python/iqm: fix subscribing to neighbours
During the conversion to use the IoT3 Core SDK in commit 7960edda5d2a
(python/iqm: use iot3.core.mqtt for queue-manager), a regression was
introduced when subscribing to the neighbours: the topic of the
interqueue was subscribed to on the local queue manager rather than on
the neighbour.

Fix this by properly subscribing to the neighbour.

Fixes: #130

Reported-by: Frédéric GARDES <frederic.gardes@orange.com>
Signed-off-by: Yann E. MORIN <yann.morin@orange.com>