python/iot3/mobility: add base ETSI message
Ideally, we would like to have message_from_json() be defined as an
static method in the etsi.Message class, but that is not possible:
when we eventually add a new class for an actual message (say, CAM),
that class will inherit from etsi.Message, while at the same time
etsi.Message will need to instanciate a CAM; so CAM will need the etsi
module to be fully imported, while at the same time, the etsi module
would need the CAM module to be fully imported. This ia checken-n-egg
problem that is not easily (if at all!) solved in python (in C/C++, we'd
have forward declaration to solve that, but it's not possible in
python).

So, implement that as a module-level function.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>