python/iqm: explicitly disable TLS for neighbours
As explained in b9c9fea83242 (python/iqm: add support for MQTT authority
over TLS), TLS is explicitly not supported for neighbours, and will have
to be done in a future evolution when accoiunting for #168.

What b9c9fea83242 forgot, though was to explicitly disable TLS, which
implicitly means that only port 1883 is supported for neighbours as this
is the only port where the default is to not enable TLS.

Fix that by explcitly disabling TLS for neighbours.

Fixes: #194

Reported-by: Frédéric GARDES <frederic.gardes@orange.com>
Signed-off-by: Yann E. MORIN <yann.morin@orange.com>