python/vehicle: fix propagating messages to local broker
Commit c482cb16774f (python/vehicle: use iot3.core.mqtt) introduced a
regression in the propagation of messages from the remote broker to
the local one.

The API for the IoT3 Core SDK extensively uses keywords rather than
positional parameters, and although this was properly accounted when
publishing messages on the remote broker, this was overlooked when
publishing messages on the local broker.

Such an issue would have raised an exception, and we would have caught
it quite easily. Alas, in an attempt at being robust when decoding the
received messages, we decided for a blanket coverage of all exceptions,
no matter what. This indeed hid the TypeError exception that is raised
when improperly calling the SDK...

So, now, we could just catch KeyError when looking up the source_uuid
field, but that would assume the message was decoded as a dict, which we
have no guarantee it was. If that were not the case, i.e.e the message
is not an ITS message, and contains any other JSON object that is not a
dictionary, that would _also_ raise TypeError, so we would have to catch
it anyway to ignore it. And we would not have been able to notice the
TypeError raised by improperly calling the SDK either anyway...

Sigh...

For now, just fix calling the SDK; we'll handle the rest later (Famous
Last Words™)...

Fixes: #203

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>