python/its-vehicle: make authentication configurable
Commit 863ecb59aa4b (python/vehicle: hot-fix for OpenTelemetry
authentication) hard-coded the authentication for the purpose of
a release, which was acceptable as its-vehicle will always use
authentication in practice.

However, we want to be able to disable authentication for local
tests, or to use Digest when it becomes available.

Make the authentication scheme configurable, like username and
password are.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>