github/workflows: add python IoT3 SDK
This new workflows differs from the existing ones, for two reasons:

 1. the IoT3 SDK uses the "new" PEP517 packaging;
 2. we want reproducible tooling, notably for black.

The other workflows use an unpinned version of black, which leads to new
issues being reported when new versions of black get released. This is
causing issues because black 20.8b1, which we uses locally, reports less
issues than black 24.10.0 that is current and now gets used in the
workflows.

So we pin all the versions of the tools weneed: black, build, and twine.
We also pin all their dependencies. Notably, pkginfo needs to be
specifically pinned to 1.10.0, as one tool requires >=1.8.0, thus 1.12.0
is installed (being the current latest version), but blck requires
pkginfo <1.11.0, so black fails to install. We pick pkginfo 1.10.0 as
the highest versions supported by both.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>