python/iot3/mobility: fix coding style
Older versions of black would not warn about those constructs, but newer
ones do. Our workflows are using unpinned versions of the tools, of
which black, so we now hit the issue (also, we did not have a workflow
for iot3 either, but black was run locally).

So, fix the warnings as black 24.10.0 reports, except for the abstract
Message.__init__(), which we fix by providing a doc-string, like is done
for ETSI.__init__() above.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>