python/iot3: use a more recent its-quadkeys
This will allow us to align the various versions we use when
we migrate the other python packages to use PEP517.

We can't (easily) use packages relative to the current one, however,
so we have to use the git repo until we eventually publish this to a
public repo (like Pypi), when we'll be able to fallback to using
actual versions. But we're not there yet.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>