python/iot3/core: simplify OTLP config in simple API
We currently use a loop to iterate over all the possible compression and
authentication schemes, and compare their string representation against
the valude from the configuration, which is a string.

However, python's enum.strEnum() constructor does accept a string as a
parameter, where that string is the str() representation of one of the
enum values, which incidentally also plays the role of validating the
value from the configuration.

Use that, rather than our canned, ugly for-loop.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>