java/iot3mobility: replace all calls to JSON isEmpty() method
The recent JSON library used in this Java library project suggests using isEmpty() to check for empty JSONObjects rather than checking that their length == 0.
However, this creates an incompatibility of this library with Android apps, because Android embeds an older JSON library version which does not have the isEmpty() method, and it does not seem possible to override this embedded JSON library with a more modern version.

A new isNullOrEmpty(...) method for JSONObject and JSONArray has been added to the JsonUtil class as a workaround, testing emptiness with a zero length check.

Signed-off-by: Mathieu LEFEBVRE <mathieu1.lefebvre@orange.com>