swift: add MQTT reconnection mechanism
The reconnection mechanism tries to reconnect to the server until it's done or an external disconnection is performed.
A list of subscriptions is managed to resume subscriptions after reconnection.
If a subcription/unsubscription is done when reconnecting, the list is updated to resume the current subscriptions when connected.
If a message is published when reconnecting, the message is ignored.

Signed-off-by: François Suc <francois.suc@orange.com>