python/requirements: update twine
Workflows have recently started to fail because of a twine error [0],
even though there had been no change in the python code itself ([0] was
a Swift chamnge). There have been similar reports elsewhere [1], and the
suggestion was to update the pkinfo dependency.

Twine 5.1.1 only accepts pkginfo up to 1.10.x, which we already use, so
to update pkginfo, we need a newer twine.

Update twine and pkginfo to he latest versions available as of today.
Add a new twine dependency as well.

With those versions, there are no error reported anymore...

[0] https://github.com/Orange-OpenSource/its-client/actions/runs/13964634913/job/39092188302
[1] https://github.com/pypi/warehouse/issues/15611#issuecomment-2003569493

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>