github/workflows: pin image version used for python workflows
We want the workflows to be reproducible, Indeed, when one pushes a new
branch that does not affect a specific language, we do not want the
workflows for that language to start failing. This however occurs when
the build tools versions change, e.g. when the distro changes. Using the
`-latest` tag ensures this failure will happen; this is unfortuante.

Switch to using a pinned version of the image, so that we have a sane
ground to start building on.

The actions/runner-images version evolve quite frequently; for example,
there are two versions tagged about a week apart 20240609.1 [0] and
20250316.1 [1]. However, it is not possible to explicitly request one
or the other; only the "main" version of the image is available (here,
24.04). This is not enough to guarantee reproducibility (e.g we've had
such issues already [2]), but is the best we can get for now.

[0] https://github.com/actions/runner-images/releases/tag/ubuntu24/20250309.1
[1] https://github.com/actions/runner-images/releases/tag/ubuntu24/20250316.1
[2] https://github.com/ymorin-orange/its-client/actions/runs/14053809854/job/39348895194

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>