python/iot3/mobility: fix setting DENM cause
The DENM cause, like all numbers in ETSI messages, is an integer, while
the parameter we expect is a higher-level object DENM.Cause enum.

Store the enum value in the ETSI message.

Properly catching this would require testing the types, as we use
enum.IntEnum that for all practical purposes act and look like ints.
This would be a much bigger endeavour to do it right...

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>