python/iot3/mobility: add sub-cause to DENM
The following commands were used to establish the initial skeleton of
the enum classes and the type alias:

  - SubCause enums:
    $ cat ../../schema/denm/denm_schema_2-1-0.json \
        |jq -r .properties.message.properties.situation_container.properties.event_type.properties.subcause.description \
        |sed -r -e 's/   +//g' \
        |sed -r -e 's/[[:space:]]//g; s/\<([^[:space:]:.,]+):/\n        class \1SubCause(enum.IntEnum):\n/g;
                    s/\<([^[:space:]().,]+)\(([[:digit:]]+)\)([.,]|$)/            \1 = \2\n/g;
                    s/-/_/g'

  - TypeAlias:
    $ cat ../../schema/denm/denm_schema_2-1-0.json \
        |jq -r .properties.message.properties.situation_container.properties.event_type.properties.subcause.description \
        |sed -r -e 's/   +//g' \
        |sed -r -e 's/[[:space:]]//g;
                    s/\<([^[:space:]:.,]+):/ SubCauses.\1SubCause |\n/g;
                    s/\<([^[:space:]().,]+)\(([[:digit:]]+)\)([.,]|$)//g;
                    s/-/_/g'

The getter for the sub-clause is a bit convoluted, because we need to
use the sub-clause type that matches the cause; i.e. we need to use the
SubCause.TrafficCondition type for the Cause.trafficCondition enum case.

The setter is a bit simpler, as we're only interested in the actual value
there; if the value is None, we need to delete the subcause from the
message.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>