python/iot3/mobility: fix setting DENM termination
The DENM termination, like all numbers in ETSI messages, is an integer,
while the parameter we expect is a higher-level object
DENM.TerminationType enum.

Store the enum value in the ETSI message.

Note that the test test-iot3-mobility-message already civered that
situation, but we use enum.IntEnum which are for all practical purposes
just an int, so they can get serialised into json or get re-used to
initialise another such enum, so we could not catch the discrepancy.

Properly catching this would require testing the types, but this is a
much bigger endeavour...

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>