github: add trampoline to select a PR template
When there are multiple PR templates in a repository, GitHub does not
offer a pull-down menu to select one. This is a known issue that annoys
people [0].

Do as suggested in a reply [1] to that thread: add a generic PR template
that lists indivdual templates that users can select from a list; this
is not very user-friendly, but there is no better known way for now...

While at it, add a little blurb of our own for those components which we
do not have tempates for yet.

[0] https://stackoverflow.com/q/73771068
[1] https://stackoverflow.com/a/75030350

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>