rust: fix channel closing propagation
When channels were switched from mspc to crossbeam it seems that sender
dropping was not correctly catched on receiver side.

It seems that the `for element in receiver` does not break the loop when
the sender is empty and disconnected.
Use a custom loop and the blocking `recv()` method to catch the
sender's disconnection

Signed-off-by: Nicolas Buffon <nicolas.buffon@orange.com>