rust: JSON reformatting for unit tests
Strangely very old tests started not to pass anymore as JSON content is
tested for equality but JSON fields order not being guaranteed in any
way two tests were failing.
Using the Rust output as expected input (only the formatting indeed not
the values) for the tests to pass again.

Signed-off-by: Nicolas Buffon <nicolas.buffon@orange.com>