java/iot3core: add topic management mechanism to the MQTT client
The MQTT client now keeps track of desired topics that it should resubscribe to in case of re-connection.

It also keeps track of topics being in the process of subscribe or unsubscribe.

In case a topic is no longer desired and its unsubscribe failed, the messages from this topic are filtered.

Signed-off-by: Mathieu LEFEBVRE <mathieu1.lefebvre@orange.com>