java/iot3mobility: remove LwM2M LocationUpdate in the sendPosition(...) method
This method wasn't checking if the LwM2M client was instantiated, causing an error before the CAM could be sent. It should anyway not be called at 1Hz, which is a much higher frequency than what we intend to use LwM2M for.

LwM2M LocationUpdates will have to be handled elsewhere and with other triggers.

Signed-off-by: Mathieu LEFEBVRE <mathieu1.lefebvre@orange.com>