python/iot3/mobility: add TAI<->UTC conversion
The ETSI time is based on TAI, not UTC, but all a UNIX-like computer
provides is UTC-based time.

Add a module that can do the back-n-forth between TAI and UTC.

Note that there is a limitation when going from UTC to TAI *while a
leapsecond is occurring*: during that time, there is no way from UTC to
know whether the leapsecond is scheduled or on-going, which means that
there is a 1-second window during which accurate conversion from UTC to
TAI is not possible. The other way around is always correct, though
(there are 2 seconds in TAI that match to the same second in UTC).

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>