python/iot3/mobility: reject leapseconds DB that has expired
When the list of leap-seconds has expired, there is no way to know
whether any new leapsecond ahs been introduced, and thus there is no wau
to convert to/from TAI anymore. Thus, we must reject any list of
leap-seconds is expired (something must be responsible for keeping it up
to date)...

Update the PR template to update the list before launching the test
suite.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>