python/iqm: add support for using MQTT shared subscriptions
This IQM implementation is merely a proof-of-concept, to deomnstrate
the behaviour of an ITS system; it was never thought to be deplyed at
scale.

As such, it was written in python for a "quick" prototyping phase.
It turns out it is working relatively nicely, but it obviously does not
scale beyond a few thousands messages ingressed per second.

However, there is no state in handling the ITS messages. Instead of
trying to scale vertically, we can very well scale horizontally and
spread the load over multiple IQMs running on the same broker, with the
broker delivering only a fraction of the toaly ingressed messages.

This is doable by leveraging the MQTT shared subscription mechanism.

To support such as setup. the IQM will have to subscribe to a specially
formatted topic, and all the IQMs instances sharing the load will have
to use a common "share name" so that the broker can recognise the
sharing and spread the load across the corresponding IQMs. We use the
instance ID as the share name.

As usual, the client-ID must be unique per MQTT session, though, so we
expand it with a long-enough random string to decrease the probablility
of a collision, should multiple IQMs share the same configuration file.

We do that for both the local broker, to be able to ingress the local
load, but also for the neighbouring brokers, to ingress the load from
there as well.

It is the responsibility of an upper management layer to spawn as many
IQMs as it needs, though; it can be statically assigned (e.g. with
docker compose) or dynamically sclaed (e.g. with k8s). The IQM does not
catter to that management part at all.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>