python/iot3/otel: create a bounded queue to pass spans
Currently, the SimpleQueue used to queue spans from the application, to
be de-queued by the sender thread, is unbounded (SimpleQueues are always
unbounded).

This means that, if the application is able to queue more spans than we
can send to the collector, the queue never fills (it's unbounded), and
spans are accumulated in huge numbers.

This happens, for example, when the telemetry server is unreachable:
when we have accumulated 'max_backlog' spans, we try to send them to the
OTLP collector; as it is unreachable, we keep the spans around, and on
the next span, we again try to send them to the collector, to fail
again, so we keep the spans, and so on.

Trying to send the spans to the collector is expensive: it requires
resolving a hostname (which may itself needs doing a network connection
to a remote host), then, if the hostname could be resolved, trying to
establish a TCP connection, which ultimately fails after the timeout of
the TCP handshake (30s).

And then we'd try again for the next span to be dequeued. And over and
over again...

Since the queue is unbounded, two cases may happen: the memory is
exhausted by an ever growing queue, or the application stops queuing
spans (for whatever reason) and the queue is slowly emptied one span at
a time (which can take quite a looong time...).

To alleviate the issue, we switch to using a Queue (instead of a
SimpleQueue). A Queue can be bounded, and we do so by accepting to queue
twice the max_backlog (crude heuristic so that the application can still
queue spans while the sender thread has not had time to dequeue a span
yet).

Note that this implies that queueing a span may block the application
while the queue is full; this situation can last for as long it takes
the sender thread to (try and fail to) send spans to the OTLP collector.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>