python/iot3/otel: do not block on full queue when sending span
Currently, when the queue to batch spans is full, the application
sending spans gets blocked until there is room for sending the spans.

This usually is a short-lived situation, but can take an inordinate
amount of time when the OTLP collector is unreachable.

Blocking the application is not nice (at all), so better lose telemetry
rather than degrade the operation of the application.

Note: when we have support for OTLP events (in addition to traces), we
can raise an event to track the situation: that's for another day,
though.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>