python/iot3/otel: simplify loop waiting for spans
When doing time-based sending, we are trying to detect the case where we
were scheduled out for longer than the batch period, and if so, send the
queued spans right away.

However, this suffers from two flaws: first, this does not account for
the case where we were suspended for *very* long (i.e. more than one
batch_period) and second, the code is cryptic.

We can do without that heuristic, though: if the batch_period is small,
we'll soon get another opportunity to send the spans; if it is large,
then the probability to be suspended that long is very slim (in which
case, too bad).

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>