python/iot3/otel: auth and conpression are not an Optional[]
Soecifying the authentication is optional (i.e. caller does not need to
specify it, in which case there will be no authentication), but it is
not an Optional[]: it can't be 'None' (the python 'None' value), as no
authentication is 'Auth.NONE'.

Ditto for compression, which is a Compression that can be the special
value 'Compression.NONE'.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>