python/iot3/otel: always do a max_backlog-based sending
When doing a batch_period-only sending, i.e. with max_backlog == None,
when the OTLP collector is not reachable, we end up never limiting the
backlog of spans to send, always keeping all spans around.

If the OTLP collector is never reachable, we forever consume more and
more memory (this can't end well).

Simplify the situation by always doing a backlog-based sending. This
ensures that we eventually stabilise our memory consumption.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>