Revert "python/iqm: use only UNIX socket to connect to local MQTT broker"
The rationale behind 9b2c91ce434b (python/iqm: use only UNIX socket to
connect to local MQTT broker) has proven to become wrong since then: now
that we can spawn multiple IQMs to share the load, it will start to make
sense to have the IQMs running on different machines, and this may no
longer be co-located on the same machine as the broker.

Re-instate the possibility to connect to the broker via TCP; still keep
the UNIX socket for simpler setups, or for machines beefy enough to host
multiple IQMs.

This reverts commit 9b2c91ce434b251a97f32b8d6f47fd487b5c5041.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>