python/iqm: add helper to convert "boolean" string to bool
configparser.ConfigParser() only knows about strings, and does not
handle anything else, especialy not None of bool. However, we do
use None in a few places, so we convert the ConfigParser() to a dict
before we use it everywhere.

This means that we no longer have access to the ConfigParser.getbool().

Add a helper that allows easily converting a string-with-a-boolean into
an actual boolean.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>