python/iqm: fix using TLS with MQTT authority
If the tls key is not provided, we end up with None, which makes for
automatic TLS setup, which is exactly what we want.

But if the tls key *is* provided, it is a string, not a boolean, as it
comes from a configparser.ConfigParser(), which only handles strings;
when converted to a dict(), there is no type conversion, so we still get
a string.

A non-empty string always tests True, whatever its content. So the
string "False" tests as True, so we end up always using TLS, even when
explicitly instructed not to.

Fix that by using the newly introduced helper, that converts a
boolean-in-a-string into a bool().

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>